Fix circular import by moving AttachConversationMiddleware to separate module
